refactor: fixed check all
