+++ components/active/documentation/IntegrationsList.vue
@@ -26,8 +26,8 @@
 </template>
 
 <script setup lang="ts">
-  import { useIntegrationsList } from '~/store/documentation/integrationsList';
   import { useI18n } from 'vue-i18n';
+  import { useIntegrationsList } from '~/composables/store/useIntegrationList';
 
   const { t } = useI18n();
   const { integrations, filterParams, refresh } = useIntegrationsList();
